Instantiate Riemann solver from factory
